requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
accelerate>=0.25.0
qwen-vl-utils

# Vector DB
qdrant-client>=1.7.0

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Video/Image Processing
opencv-python>=4.9.0
Pillow>=10.0.0

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
tqdm>=4.66.0
numpy>=1.24.0
requests>=2.31.0

# Message Queue
pika>=1.3.0

# AWS S3
boto3>=1.34.0

# Geospatial
pymongo>=4.6.0
shapely>=2.0.0
dnspython>=2.6.0
pyproj>=3.6.0

# Observability
langfuse>=3.0.0

# Development / Load Testing
pytest>=7.4.0
httpx>=0.25.0
aiohttp>=3.9.0
